requirements.txt
strawberry-graphql[fastapi]==0.246.1
fastapi[all]==0.110.1
sqlalchemy==2.0.16
psycopg2-binary==2.9.6
uvicorn==0.23.2
